+++ java/BUILD
@@ -1,0 +1,7 @@
+load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
+
+bzl_library(
+    name = "bzl",
+    srcs = ["java.bzl"],
+    visibility = ["//:internal"],
+)
