+++ java/README.md
@@ -1,5 +1,9 @@
+# Bazel Integration Testing - Java
+
 ## Usage
+
 ### BUILD.bazel
+
 ```python
 load("//:bazel_integration_test.bzl", "bazel_java_integration_test")
 bazel_java_integration_test(
@@ -7,7 +11,10 @@ bazel_java_integration_test(
     srcs = ["MyTest.java"],
 )
 ```
-If you want your test to be in a different JVM language you can use the following:
+
+If you want your test to be in a different JVM language you can use the
+following:
+
 ```python
 bazel_java_integration_test(
     name = "MyScalaTest",
@@ -15,10 +22,15 @@ bazel_java_integration_test(
     runtime_deps = [":MyScalaTestSrc"],
 )
 ```
-Where `MyScalaTestSrc` is a `scala_library` you add yourself and has `com.example.tests.integration.MyTestFromScala` class.       
+
+Where `MyScalaTestSrc` is a `scala_library` you add yourself and has
+`com.example.tests.integration.MyTestFromScala` class.
+
 ### Test
-**Note** your test class must be run by JUnit as its run by `java_test`.  
-Very simplistic example:
+
+**Note** your test class must be run by JUnit as its run by `java_test`.  A
+very simple example:
+
 ```java
 public class ExampleTest {
   private WorkspaceDriver driver = new WorkspaceDriver();
@@ -49,7 +61,18 @@ public class ExampleTest {
   }
 }
 ```  
-In your test you should work with the `WorkspaceDriver` which has many useful utility methods for working with scratch workspaces.  
-These include, but are not limited to, `copyFromRunfiles`/`scratchFile`/`workspaceDirectoryContents`(useful for verbose error messages) and of course `bazelCommand` which allows you to run `bazel build` / `bazel test` / `bazel query` and more.  
-The [WorkspaceDriverIntegrationTest](../javatests/build/bazel/tests/integration/WorkspaceDriverIntegrationTest.java) has a few examples of realistic tests which sets up the environment and asserts Bazel's outputs (exit code, stderr, etc).  
-[WorkspaceDriverTest](../javatests/build/bazel/tests/integration/WorkspaceDriverTest.java) and [BazelBaseTestCaseTest](../javatests/build/bazel/tests/integration/BazelBaseTestCaseTest.java) show some more of the features.
+
+In your test you should work with the
+[`WorkspaceDriver`](build/bazel/tests/integration/WorkspaceDriver.java) which
+has many useful utility methods for working with scratch workspaces. These
+include, but are not limited to, `copyFromRunfiles`, `scratchFile`,
+`workspaceDirectoryContents` (useful for verbose error messages), and of course
+`bazelCommand` which allows you to run `bazel build`, `bazel test`, `bazel
+query`, and more.  The
+[WorkspaceDriverIntegrationTest](../javatests/build/bazel/tests/integration/WorkspaceDriverIntegrationTest.java)
+has a few examples of realistic tests which sets up the environment and asserts
+Bazel's outputs (exit code, stderr, etc).
+[WorkspaceDriverTest](../javatests/build/bazel/tests/integration/WorkspaceDriverTest.java)
+and
+[BazelBaseTestCaseTest](../javatests/build/bazel/tests/integration/BazelBaseTestCaseTest.java)
+show some more of the features.
